# silvera-solution/scipaper.io/output/SciPaper/openapi.json
{"openapi": "3.0.0", "info": {"version": "0.0.1b", "title": "SciPaper", "description": ""}, "servers": [{"url": "http://localhost:9095/papers"}], "paths": {"/paper": {"post": {"description": "Creates a new paper.", "operationId": "createPaper", "requestBody": {"description": "Paper to add.", "required": true, "content": {"application/json": {"schema": {"$ref": "#components/schemas/Paper"}}}}, "responses": {"200": {"$ref": "#components/responses/PaperOK"}}}}, "/paper/id": {"put": {"description": "Update Paper object.", "operationId": "updatePaper", "parameters": [{"name": "id", "in": "path", "description": "ID of paper to update", "required": true, "schema": {"type": "string"}}], "requestBody": {"$ref": "#components/requestBodies/PaperBody"}, "responses": {"200": {"$ref": "#components/responses/PaperOK"}}}, "get": {"description": "Read Paper object.", "operationId": "readPaper", "parameters": [{"name": "id", "in": "path", "description": "ID of paper to update", "required": true, "schema": {"type": "string"}}], "requestBody": {"$ref": "#components/requestBodies/PaperBody"}, "responses": {"200": {"$ref": "#components/responses/PaperOK"}}}, "delete": {"description": "Delete Paper object.", "operationId": "deletePaper", "parameters": [{"name": "id", "in": "path", "description": "ID of paper to update", "required": true, "schema": {"type": "string"}}], "requestBody": {"$ref": "#components/requestBodies/PaperBody"}, "responses": {"200": {"$ref": "#components/responses/PaperOK"}}}}, "//my-papers": {"get": {"description": "", "operationId": "getAllByAuthor", "parameters": [{"name": "authorId", "in": "path", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "JSON representation of Paper object", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Paper"}}}}}}}}, "/publish": {"post": {"description": "", "operationId": "publish", "parameters": [{"name": "paperId", "in": "path", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "", "content": {}}}}}}, "components": {"schemas": {"Paper": {"type": "object", "properties": {"title": {"type": "string"}, "author": {"type": "string"}, "authorId": {"type": "string"}, "sections": {"type": "array", "items": {"$ref": "#/components/schemas/Section"}}}, "required": ["title", "sections"]}, "Section": {"type": "object", "properties": {"name": {"type": "string"}, "content": {"type": "string"}}, "required": ["name", "content"]}}, "requestBodies": {"PaperBody": {"description": "JSON representation of Paper object", "required": true, "content": {"application/json": {"schema": {"$ref": "#components/schemas/Paper"}}}}, "SectionBody": {"description": "JSON representation of Section object", "required": true, "content": {"application/json": {"schema": {"$ref": "#components/schemas/Section"}}}}}, "responses": {"PaperOK": {"description": "JSON representation of Paper object", "content": {"application/json": {"schema": {"$ref": "#components/schemas/Paper"}}}}, "SectionOK": {"description": "JSON representation of Section object", "content": {"application/json": {"schema": {"$ref": "#components/schemas/Section"}}}}}}}