# silvera-solution/scipaper.io/output/SciPaper/src/main/java/com/silvera/SciPaper/controller/SciPaperController.java
/**
    THIS IS GENERATED CODE AND SHOULD NOT BE CHANGED MANUALLY!!!

    Generated by: silvera
    Date: 2022-05-19 16:57:01
*/

package com.silvera.SciPaper.controller;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.cloud.context.config.annotation.RefreshScope;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.ResponseBody;
import com.silvera.SciPaper.service.base.*;
import com.silvera.SciPaper.service.impl.SciPaperService;
import com.silvera.SciPaper.domain.model.*;
import com.silvera.SciPaper.dto.PublishRequest;

import org.springframework.web.bind.annotation.*;

import javax.servlet.http.HttpServletRequest;
import javax.validation.Valid;


@RefreshScope
@RestController
public class SciPaperController {

    @Autowired
    ISciPaperService sciPaperService;

    // Auto-generated CRUD methods
    
    @RequestMapping(value="/create-update", method=RequestMethod.POST)
    @ResponseBody
    public Paper createPaper(@Valid @RequestBody Paper paper, HttpServletRequest request){
        ((SciPaperService)sciPaperService).checkAuthorized(request);
        ((SciPaperService)sciPaperService).setAuthorForPaper(request, paper);
        return sciPaperService.createPaper(paper);
    }
    
    
    @RequestMapping(value="/create-update", method=RequestMethod.PUT)
    @ResponseBody
    public Paper updatePaper(@Valid @RequestBody Paper paper, HttpServletRequest request){
        ((SciPaperService)sciPaperService).checkAuthorized(request);
        ((SciPaperService)sciPaperService).setAuthorForPaper(request, paper);
        return sciPaperService.updatePaper(paper.getId(), paper);
    }
    
    
    @RequestMapping(value="/paper/{id}", method=RequestMethod.GET)
    @ResponseBody
    public Paper readPaper(@PathVariable java.lang.String id, HttpServletRequest request){
        ((SciPaperService)sciPaperService).checkAuthorized(request);
        return sciPaperService.readPaper(id);
    }
    
    
    // @RequestMapping(value="/paper/{id}", method=RequestMethod.DELETE)
    // @ResponseBody
    // public void deletePaper(@PathVariable java.lang.String id){
    //     sciPaperService.deletePaper(id);
    // }
    

    //
    // Public functions
    //
    
    @GetMapping(value="/my-papers")
    public java.util.List<Paper> getAllByAuthor(HttpServletRequest request){
        return sciPaperService.getAllByAuthor(((SciPaperService)sciPaperService).getUsername(request));
    }
    

    @PostMapping(value = "/publish")
    public void publish(@RequestBody PublishRequest pr, HttpServletRequest request){
         sciPaperService.publish(pr.paperId, ((SciPaperService)sciPaperService).getUsername(request));
    }
    
}