# silvera-solution/scipaper.io/output/SciPaper/src/main/java/com/silvera/SciPaper/domain/model/Paper.java
/**
    THIS IS GENERATED CODE AND SHOULD NOT BE CHANGED MANUALLY!!!

    Generated by: silvera
    Date: 2022-05-19 22:55:47
*/

package com.silvera.SciPaper.domain.model;

import org.springframework.data.annotation.Id;
import javax.validation.constraints.*;
public class Paper {

    
    @Id
    private String id;
    
    
    
    @NotBlank(message="Field 'title' cannot be empty!")
    private java.lang.String title;
    
    private java.lang.String author;
    
    private java.lang.String authorId;
    
    
    @NotNull(message="Field 'sections' is mandatory!")
    private java.util.List<Section> sections= java.util.Collections.emptyList();


    
    public java.lang.String getTitle() {
        return this.title;
    }

    public void setTitle(java.lang.String title) {
        this.title = title;
    }
    
    public java.lang.String getAuthor() {
        return this.author;
    }

    public void setAuthor(java.lang.String author) {
        this.author = author;
    }
    
    public java.lang.String getAuthorId() {
        return this.authorId;
    }

    public void setAuthorId(java.lang.String authorId) {
        this.authorId = authorId;
    }
    
    public java.util.List<Section> getSections() {
        return this.sections;
    }

    public void setSections(java.util.List<Section> sections) {
        this.sections = sections;
    }
    

    
    public String getId(){
        return this.id;
    }
    

}